+++ src/net/lodoma/lime/gui/GUIElement.java
@@ -7,6 +7,6 @@ public interface GUIElement
     public void create();
     public void destroy();
     
-    public void update(float timeDelta, Vector2 mousePosition);
+    public void update(double timeDelta, Vector2 mousePosition);
     public void render();
 }
